noq

Module congestion

Logic for controlling the rate at which data is sent

Structs

Bbr3

Experimental! Use at your own risk.

Bbr3Config

Configuration for the Bbr3 congestion controller Different pacing_gains can be set to modify the multiplier used to increase the sending rates. Different cwnd_gains can be set to modify the multiplier used to increase the congestion windows. All of these parameters are specific to different states of the algorithm: see BbrState pacing_margin_percent is used to set a margin when calculating the pacing_rate in order to not send at 100% capacity when calculating pacing.

ControllerMetrics

Common congestion controller metrics used both for logging purposes but also to alter the pacing of the connection with pacing_rate and send_quantum

Cubic

The RFC8312 congestion controller, as widely used for TCP

CubicConfig

Configuration for the Cubic congestion controller

NewReno

A simple, standard congestion controller

NewRenoConfig

Configuration for the NewReno congestion controller

Traits

Controller

Common interface for different congestion controllers

ControllerFactory

Constructs controllers on demand