noq

Struct Path

pub struct Path { /* private fields */ }

An open network transmission within a multipath-enabled connection.

As long as a Path or WeakPathHandle is alive, it is ensured that the PathStats for this path are not dropped even after the path is abandoned.

Implementations

impl Path

pub fn weak_handle(&self) -> WeakPathHandle

Returns a WeakPathHandle for this path.

Holding a WeakPathHandle does not keep a connection alive, but ensures that the path’s stats are not dropped until the underlying connection is dropped, even if the path is abandoned.

pub fn id(&self) -> PathId

The PathId of this path.

pub fn status(&self) -> Result<PathStatus, ClosedPath>

The current local PathStatus of this path.

pub fn set_status( &self, status: PathStatus, ) -> Result<PathStatus, SetPathStatusError>

Sets the PathStatus of this path.

Returns the previous status of the path.

pub fn stats(&self) -> PathStats

Returns the PathStats for this path.

pub fn close(&self) -> Result<(), ClosePathError>

Closes this path.

The path is immediately considered closed by the local endpoint. Once the state is removed, after a short period of time for any in-flight packets, a PathEvent::Abandoned is returned.

pub fn set_max_idle_timeout( &self, timeout: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the max idle timeout for a specific path

See TransportConfig::default_path_max_idle_timeout for details.

Returns the previous value of the setting.

pub fn set_keep_alive_interval( &self, interval: Option<Duration>, ) -> Result<Option<Duration>, ClosedPath>

Sets the keep_alive_interval for a specific path

See TransportConfig::default_path_keep_alive_interval for details.

Returns the previous value of the setting.

pub fn observed_external_addr(&self) -> Result<AddressDiscovery, ClosedPath>

Track changes on our external address as reported by the peer.

If the address-discovery extension is not negotiated, the stream will never return.

pub fn remote_address(&self) -> Result<SocketAddr, ClosedPath>

The peer’s UDP address for this path.

pub fn local_ip(&self) -> Result<Option<IpAddr>, ClosedPath>

The local IP used for this path, if known.

Returns Ok(None) for clients or when the platform does not expose this information; see noq_udp::RecvMeta::dst_ip for supported platforms.

pub fn network_path(&self) -> Result<FourTuple, ClosedPath>

The network path used for this path.

Returns a FourTuple, combining Self::remote_address and Self::local_ip.

pub fn ping(&self) -> Result<(), ClosedPath>

Ping the remote endpoint over this path.